+++ SonicTheHedgehog/Components/UniqueSkinEffect.cs
@@ -5,6 +5,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
+using UnityEngine.AddressableAssets;
+using static SonicTheHedgehog.Components.UniqueSkinEffect;
 
 namespace SonicTheHedgehog.Components
 {
@@ -16,6 +18,10 @@ namespace SonicTheHedgehog.Components
 
         private static List<string> flyingAnimationSkinTokens = new List<string>();
         public static bool[] flyingAnimationSkins;
+
+        private static Dictionary<string, SuperGrandSlamMeshReplacements> superGrandSlamMeshReplacementTokens = new Dictionary<string, SuperGrandSlamMeshReplacements>();
+        public static SuperGrandSlamMeshReplacements[] superGrandSlamMeshReplacements;
+
 
         private void Start()
         {
@@ -42,12 +48,15 @@ namespace SonicTheHedgehog.Components
             RoR2.UI.MainMenu.MainMenuController.OnMainMenuInitialised -= UniqueSkinEffect.Bake;
             SkinDef[] skins = SkinCatalog.GetBodySkinDefs(BodyCatalog.FindBodyIndex("SonicTheHedgehog"));
             flyingAnimationSkins = new bool[skins.Length];
+            superGrandSlamMeshReplacements = new SuperGrandSlamMeshReplacements[skins.Length];
             for (int i = 0; i < skins.Length; i++)
             {
                 flyingAnimationSkins[i] = flyingAnimationSkinTokens.Contains(skins[i].nameToken);
+                if (superGrandSlamMeshReplacementTokens.TryGetValue(skins[i].nameToken, out var meshReplacement)) superGrandSlamMeshReplacements[i] = meshReplacement;
             }
 
             flyingAnimationSkinTokens = null;
+            superGrandSlamMeshReplacementTokens = null;
         }
 
         public static void AddFlyingSkin(string nameToken)
@@ -58,6 +67,21 @@ namespace SonicTheHedgehog.Components
         public static void AddFlyingSkin(List<string> nameTokens)
         {
             flyingAnimationSkinTokens.Concat(nameTokens);
+        }
+
+        public static void AddSuperGrandSlamMeshReplacement(string nameToken, Mesh mesh)
+        {
+            superGrandSlamMeshReplacementTokens.Add(nameToken, new SuperGrandSlamMeshReplacements { mesh = mesh });
+        }
+        public static void AddSuperGrandSlamMeshReplacement(string nameToken, AssetReferenceT<Mesh> meshAddress)
+        {
+            superGrandSlamMeshReplacementTokens.Add(nameToken, new SuperGrandSlamMeshReplacements { meshAddress = meshAddress });
+        }
+
+        public struct SuperGrandSlamMeshReplacements
+        {
+            public Mesh mesh;
+            public AssetReferenceT<Mesh> meshAddress;
         }
     }
 }