+++ SonicTheHedgehog/Content/Achievements/SonicHomingAttackAirborneAchievement.cs
@@ -8,9 +8,11 @@ using UnityEngine;
 namespace SonicTheHedgehog.Modules.Achievements
 {
     // I have no god damn clue what SonicSkills.Parry is or what a reward identifier is I just ignore it and hope it works
-    [RegisterAchievement(SonicTheHedgehogPlugin.DEVELOPER_PREFIX + "SONICPARRYUNLOCKABLE", "SonicSkills.Parry", null, 3)]
+    [RegisterAchievement(identifier, unlockableIdentifier, null, 3)]
     public class SonicHomingAttackAirborneAchievement : BaseAchievement
     {
+        public const string identifier = SonicTheHedgehogPlugin.DEVELOPER_PREFIX + "SONICPARRYUNLOCKABLE";
+        public const string unlockableIdentifier = "SonicSkills.Parry";
         // Hit countRequired different enemies with a homing attack without touching the ground (Unlocks Parry)
         private int count;
         private readonly List<HealthComponent> hitEnemies = new List<HealthComponent>();